test ProfileSelector mapStateToProps with shallow
this is better as it avoid exporting private methods
