+++ packages/profile-selector/package.json
@@ -29,6 +29,7 @@
     "@storybook/react": "3.1.3",
     "eslint": "3.19.0",
     "jest": "20.0.4",
+    "redux-mock-store": "^1.2.3",
     "storybook-addon-a11y": "3.0.0"
   }
 }
